fix: mailings with missing default credit card
