+++ packages/republik-crowdfundings/lib/scheduler/owners.js
@@ -3,6 +3,9 @@ const moment = require('moment')
 const Promise = require('bluebird')
 
 const { transformUser } = require('@orbiting/backend-modules-auth')
+const { publishMonitor } = require('@orbiting/backend-modules-republik/lib/slack')
+
+const { getDefaultPaymentSource } = require('../paymentSource')
 
 const {
   prolongBeforeDate: getProlongBeforeDate
@@ -139,7 +142,7 @@ const createBuckets = (now) => [
     payload: {
       templateName: 'membership_owner_autopay_notice'
     },
-    handler: mailings
+    handler: autoPayMailings
   },
   {
     name: 'membership_owner_autopay',
@@ -290,4 +293,35 @@ const run = async (args, context) => {
 module.exports = {
   DAYS_BEFORE_END_DATE,
   run
+}
+
+async function autoPayMailings (user, bucket, context) {
+  const pgdb = context.pgdb
+  const defaultPaymentSource = getDefaultPaymentSource(user, pgdb)
+
+  if (!defaultPaymentSource) {
+    await setAutoPayToFalse(user, pgdb)
+    return
+  }
+
+  user = { ...user, defaultPaymentSource }
+  return mailings(user, bucket, context)
+}
+
+async function setAutoPayToFalse (user, pgdb) {
+  try {
+    await publishMonitor(
+      user,
+      `setAutoPayToFalse: \`autoPay\` was set to \`false\` due to missing default credit card\n{ADMIN_FRONTEND_BASE_URL}/users/${user.id}`
+    )
+
+    await pgdb.public.memberships.update({
+      userId: user.id
+    }, {
+      autoPay: false
+    })
+  } catch (e) {
+    // swallow slack message
+    console.warn(`publish to slack failed - autoPayMailings - user id ${user.id}`)
+  }
 }
